fixed build for OS X gcc

+++ OpenEXR/exrmakepreview/makePreview.cpp
@@ -62,7 +62,7 @@ namespace {
 float
 knee (float x, float f)
 {
-    return logf (x * f + 1) / f;
+    return log (x * f + 1) / f;
 }
 
 
@@ -82,7 +82,7 @@ gamma (half h, float m)
     if (x > 1)
 	x = 1 + knee (x - 1, 0.184874f);
 
-    return (unsigned char) (clamp (powf (x, 0.4545f) * 84.66f, 0.f, 255.f));
+    return (unsigned char) (clamp (pow (x, 0.4545f) * 84.66f, 0.f, 255.f));
 }
 
 
@@ -117,7 +117,7 @@ generatePreview (const char inFileName[],
 
     float fx = (previewWidth  > 0)? (float (w - 1) / (previewWidth  - 1)): 1;
     float fy = (previewHeight > 0)? (float (h - 1) / (previewHeight - 1)): 1;
-    float m  = powf (2.f, clamp (exposure + 2.47393f, -20.f, 20.f));
+    float m  = pow (2.f, clamp (exposure + 2.47393f, -20.f, 20.f));
 
     for (int y = 0; y < previewHeight; ++y)
     {
